+++ src/libs/OptionsListUtils/index.ts
@@ -49,6 +49,7 @@ import {
     getReceiptScanFailedMessage,
     getRenamedAction,
     getReopenedMessage,
+    getReportActionActorAccountID,
     getReportActionHtml,
     getReportActionMessageText,
     getRetractedMessage,
@@ -105,7 +106,6 @@ import {
     getReimbursementDeQueuedOrCanceledActionMessage,
     getReimbursementQueuedActionMessage,
     getRejectedReportMessage,
-    getReportActionActorAccountID,
     getReportLastMessage,
     getReportNotificationPreference,
     getReportOrDraftReport,
