+++ src/libs/ReportUtils.ts
@@ -195,6 +195,7 @@ import {
     getRenamedAction,
     getReopenedMessage,
     getReportAction,
+    getReportActionActorAccountID,
     getReportActionHtml,
     getReportActionMessage as getReportActionMessageReportUtils,
     getReportActionMessageText,
@@ -11242,65 +11243,65 @@ function canLeaveChat(report: OnyxEntry<Report>, policy: OnyxEntry<Policy>, isRe
     return (isChatThread(report) && !!getReportNotificationPreference(report)) || isUserCreatedPolicyRoom(report) || isNonAdminOrOwnerOfPolicyExpenseChat(report, policy);
 }
 
-function getReportActionActorAccountID(
-    reportAction: OnyxEntry<ReportAction>,
-    iouReport: OnyxEntry<Report>,
-    report: OnyxEntry<Report>,
-    delegatePersonalDetails?: PersonalDetails | undefined | null,
-): number | undefined {
-    switch (reportAction?.actionName) {
-        case CONST.REPORT.ACTIONS.TYPE.REPORT_PREVIEW: {
-            const ownerAccountID = iouReport?.ownerAccountID ?? reportAction?.childOwnerAccountID;
-            const actorAccountID = iouReport?.managerID ?? reportAction?.childManagerAccountID;
+// function getReportActionActorAccountID(
+//     reportAction: OnyxEntry<ReportAction>,
+//     iouReport: OnyxEntry<Report>,
+//     report: OnyxEntry<Report>,
+//     delegatePersonalDetails?: PersonalDetails | undefined | null,
+// ): number | undefined {
+//     switch (reportAction?.actionName) {
+//         case CONST.REPORT.ACTIONS.TYPE.REPORT_PREVIEW: {
+//             const ownerAccountID = iouReport?.ownerAccountID ?? reportAction?.childOwnerAccountID;
+//             const actorAccountID = iouReport?.managerID ?? reportAction?.childManagerAccountID;
 
-            if (isPolicyExpenseChat(report) || delegatePersonalDetails) {
-                return ownerAccountID;
-            }
+//             if (isPolicyExpenseChat(report) || delegatePersonalDetails) {
+//                 return ownerAccountID;
+//             }
 
-            return actorAccountID;
-        }
+//             return actorAccountID;
+//         }
 
-        case CONST.REPORT.ACTIONS.TYPE.CREATED: {
-            const reportNameValuePairs = allReportNameValuePair?.[`${ONYXKEYS.COLLECTION.REPORT_NAME_VALUE_PAIRS}${iouReport?.reportID}`];
-            if (isHarvestCreatedExpenseReport(reportNameValuePairs?.origin, reportNameValuePairs?.originalID)) {
-                return CONST.ACCOUNT_ID.CONCIERGE;
-            }
-            return reportAction?.actorAccountID;
-        }
+//         case CONST.REPORT.ACTIONS.TYPE.CREATED: {
+//             const reportNameValuePairs = allReportNameValuePair?.[`${ONYXKEYS.COLLECTION.REPORT_NAME_VALUE_PAIRS}${iouReport?.reportID}`];
+//             if (isHarvestCreatedExpenseReport(reportNameValuePairs?.origin, reportNameValuePairs?.originalID)) {
+//                 return CONST.ACCOUNT_ID.CONCIERGE;
+//             }
+//             return reportAction?.actorAccountID;
+//         }
 
-        case CONST.REPORT.ACTIONS.TYPE.SUBMITTED:
-        case CONST.REPORT.ACTIONS.TYPE.SUBMITTED_AND_CLOSED:
-        case CONST.REPORT.ACTIONS.TYPE.APPROVED:
-        case CONST.REPORT.ACTIONS.TYPE.FORWARDED:
-        case CONST.REPORT.ACTIONS.TYPE.IOU: {
-            const originalMessage = getOriginalMessage(reportAction);
-            const actionName = reportAction?.actionName;
+//         case CONST.REPORT.ACTIONS.TYPE.SUBMITTED:
+//         case CONST.REPORT.ACTIONS.TYPE.SUBMITTED_AND_CLOSED:
+//         case CONST.REPORT.ACTIONS.TYPE.APPROVED:
+//         case CONST.REPORT.ACTIONS.TYPE.FORWARDED:
+//         case CONST.REPORT.ACTIONS.TYPE.IOU: {
+//             const originalMessage = getOriginalMessage(reportAction);
+//             const actionName = reportAction?.actionName;
 
-            // Check if this should show Concierge as the actor
-            const wasSubmittedViaHarvesting = originalMessage && 'harvesting' in originalMessage ? originalMessage.harvesting : false;
-            const wasAutomatic = originalMessage && 'automaticAction' in originalMessage ? originalMessage.automaticAction : false;
-            const isPayment = originalMessage && 'type' in originalMessage && originalMessage.type === CONST.IOU.REPORT_ACTION_TYPE.PAY;
+//             // Check if this should show Concierge as the actor
+//             const wasSubmittedViaHarvesting = originalMessage && 'harvesting' in originalMessage ? originalMessage.harvesting : false;
+//             const wasAutomatic = originalMessage && 'automaticAction' in originalMessage ? originalMessage.automaticAction : false;
+//             const isPayment = originalMessage && 'type' in originalMessage && originalMessage.type === CONST.IOU.REPORT_ACTION_TYPE.PAY;
 
-            // Show Concierge for:
-            // - Harvesting (delayed submissions)
-            // - Automatic approvals/forwards via workspace rules
-            // - Automatic payments via workspace rules
-            if (wasSubmittedViaHarvesting || (wasAutomatic && actionName !== CONST.REPORT.ACTIONS.TYPE.IOU) || (wasAutomatic && isPayment)) {
-                return CONST.ACCOUNT_ID.CONCIERGE;
-            }
+//             // Show Concierge for:
+//             // - Harvesting (delayed submissions)
+//             // - Automatic approvals/forwards via workspace rules
+//             // - Automatic payments via workspace rules
+//             if (wasSubmittedViaHarvesting || (wasAutomatic && actionName !== CONST.REPORT.ACTIONS.TYPE.IOU) || (wasAutomatic && isPayment)) {
+//                 return CONST.ACCOUNT_ID.CONCIERGE;
+//             }
 
-            // For SUBMITTED actions, check adminAccountID first (admin-submit case)
-            if (actionName === CONST.REPORT.ACTIONS.TYPE.SUBMITTED || actionName === CONST.REPORT.ACTIONS.TYPE.SUBMITTED_AND_CLOSED) {
-                return reportAction?.adminAccountID ?? reportAction?.actorAccountID;
-            }
+//             // For SUBMITTED actions, check adminAccountID first (admin-submit case)
+//             if (actionName === CONST.REPORT.ACTIONS.TYPE.SUBMITTED || actionName === CONST.REPORT.ACTIONS.TYPE.SUBMITTED_AND_CLOSED) {
+//                 return reportAction?.adminAccountID ?? reportAction?.actorAccountID;
+//             }
 
-            return reportAction?.actorAccountID;
-        }
+//             return reportAction?.actorAccountID;
+//         }
 
-        default:
-            return reportAction?.actorAccountID;
-    }
-}
+//         default:
+//             return reportAction?.actorAccountID;
+//     }
+// }
 
 function createDraftWorkspaceAndNavigateToConfirmationScreen(introSelected: OnyxEntry<IntroSelected>, transactionID: string, actionName: IOUAction): void {
     const isCategorizing = actionName === CONST.IOU.ACTION.CATEGORIZE;
