+++ src/libs/SidebarUtils.ts
@@ -53,6 +53,7 @@ import {
     getRenamedAction,
     getReopenedMessage,
     getReportAction,
+    getReportActionActorAccountID,
     getReportActionMessageText,
     getRetractedMessage,
     getRoomAvatarUpdatedMessage,
@@ -103,7 +104,6 @@ import {
     getMovedTransactionMessage,
     getParticipantsAccountIDsForDisplay,
     getPolicyName,
-    getReportActionActorAccountID,
     getReportDescription,
     getReportMetadata, // eslint-disable-next-line @typescript-eslint/no-deprecated
     getReportName,
